requirements.txt
flask
flask-cors
python-docx
PyMuPDF
mammoth
transformers
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
chromadb
ollama